wip: refatoraçao
