+++ src/Controller/controllerInicialPage.tsx
@@ -5,7 +5,22 @@ import {
   requestChartTSCard,
 } from "../Services/Chart/request";
 import { getYear } from "../Services/localstorage";
+import { getMonthNumber } from "../Controller/controllerGlobal";
+import color from "../Styles/colors";
+import { ChartLinesModel } from "../Components/Chart/ChartLines/chartLinesModel";
 
+
+export interface Chart {
+  year: number;
+  month: number;
+  n_registers: number;
+  n_approved: number;
+}
+
+const month = [
+  "Janeiro", "Fevereiro", "Março", "Abril", "Maio", "Junho",
+  "Julho", "Agosto", "Setembro", "Outubro", "Novembro", "Dezembro"
+];
 
 const formatDate = (date: Date) => {
   const year = date.getFullYear();
@@ -44,23 +59,58 @@ const formatDate = (date: Date) => {
 
 //todo: getInicialPageMatriculatedData
 
-const getInicialPageStatusClassesData = async (ts: number[], dates: any[]) => {
+
+const getInitialPageMatriculatedData = async (ts: number[], dates: Date[]) => {
+  if (!dates || dates.length < 2 || !dates[0] || !dates[1]) return null;
+
   const start = formatDate(dates[0]);
   const end = formatDate(dates[1]);
 
   try {
-    const response = await requestChartMatriculated(start, end, ts);
-    return response.data;
+    const response = await requestChartMatriculated(start, end, ts ?? []);
+    const data: Chart[] = response.data;
+
+    const availableMonths = Array.from(new Set(data.map((item) => item.month))).sort((a, b) => a - b);
+
+    const mesInital = dates[0]?.getMonth() ?? 0;
+    const mesFinal = dates[1]?.getMonth() ?? 0;
+
+    const labels = dates[0]
+      ? getMonthNumber(mesInital, mesFinal)
+      : availableMonths.map((m) => month[m]);
+
+    const datasets = [
+      {
+        label: "Total de Matrículas Confirmadas",
+        data: availableMonths.map((m) => {
+          const found = data.find((element) => element.month === m);
+          return found ? found.n_approved : 0;
+        }),
+        borderColor: color.blue,
+        fill: false,
+      },
+      {
+        label: "Total de Matrículas",
+        data: availableMonths.map((m) => {
+          const found = data.find((element) => element.month === m);
+          return found ? found.n_registers : 0;
+        }),
+        borderColor: color.colorCardOrange,
+        fill: false,
+      },
+    ];
+
+    return new ChartLinesModel(labels, datasets);
   } catch (error) {
-    console.error("Erro ao buscar dados do gráfico:", error);
-    return [];
+    console.error("Erro ao buscar dados do gráfico de matrículas", error);
+    return new ChartLinesModel([], []);
   }
-}
-
-
+};
 //todo: getInicialPageStatusClassesData
 
 export const getInitialPageModel = async (ts: number[], dates: any[]) => {
   const cardsData = await getInitialPageCardsData(ts, dates);
-  return new InitialPageModel(cardsData);
+  const matriculatedData = (await getInitialPageMatriculatedData(ts, dates)) ?? new ChartLinesModel([], []);
+
+  return new InitialPageModel(cardsData, matriculatedData);
 };
