+++ src/Pages/InitialPage/Components/ChartMatrticulated/type.ts
@@ -1,0 +1,6 @@
+export interface Chart {
+  year: number;
+  month: number;
+  n_registers: number;
+  n_approved: number;
+}
