+++ src/Pages/InitialPage/Components/ChartStatus/type.ts
@@ -1,0 +1,6 @@
+export interface ChartStatus {
+    social_technology_name: string;
+    t_pending: bigint;
+    t_approved: bigint;
+    t_canceled: bigint;
+  }
