+++ src/Pages/InitialPage/initialPageModel.tsx
@@ -8,6 +8,16 @@ export class InitialPageModel {
     totalUserSocialTechnologies: number;
   };
 
+  matriculatedData: {
+    labels: string[];
+    datasets: {
+      label: string;
+      data: number[];
+      borderColor: string;
+      fill: boolean;
+    }[];
+  };
+
   constructor(
     cardsData: {
       totalMeetings: number;
@@ -17,8 +27,18 @@ export class InitialPageModel {
       totalProjects: number;
       totalUserSocialTechnologies: number;
     },
+    matriculatedData: {
+      labels: string[];
+      datasets: {
+        label: string;
+        data: number[];
+        borderColor: string;
+        fill: boolean;
+      }[];
+    }
     
   ) {
     this.cardsData = cardsData;
+    this.matriculatedData = matriculatedData;
   }
 }
